fix: cosmic-settings configuration issues
