+++ pages/desktop/src/wallpaper.rs
@@ -2,6 +2,7 @@ pub use cosmic_bg_config::{Color, Config, Entry, Gradient, ScalingMode, Source};
 
 use image::RgbaImage;
 use std::{
+    borrow::Cow,
     collections::{hash_map::DefaultHasher, HashMap},
     fs::DirEntry,
     hash::{Hash, Hasher},
@@ -9,6 +10,39 @@ use std::{
     sync::Arc,
 };
 use tokio::sync::mpsc::{self, Receiver};
+
+pub const DEFAULT_COLORS: &[Color] = &[
+    Color::Single([0.580, 0.922, 0.922]),
+    Color::Single([0.000, 0.286, 0.427]),
+    Color::Single([1.000, 0.678, 0.000]),
+    Color::Single([0.282, 0.725, 0.78]),
+    Color::Single([0.333, 0.278, 0.259]),
+    Color::Single([0.969, 0.878, 0.384]),
+    Color::Single([0.063, 0.165, 0.298]),
+    Color::Single([1.000, 0.843, 0.631]),
+    Color::Single([0.976, 0.227, 0.514]),
+    Color::Single([1.000, 0.612, 0.867]),
+    Color::Single([0.812, 0.490, 1.000]),
+    Color::Single([0.835, 0.549, 1.000]),
+    Color::Single([0.243, 0.533, 1.000]),
+    Color::Single([0.584, 0.769, 0.988]),
+    Color::Gradient(Gradient {
+        colors: Cow::Borrowed(&[[1.000, 0.678, 0.000], [0.282, 0.725, 0.78]]),
+        radius: 270.0,
+    }),
+    Color::Gradient(Gradient {
+        colors: Cow::Borrowed(&[[1.000, 0.843, 0.631], [0.58, 0.922, 0.922]]),
+        radius: 270.0,
+    }),
+    Color::Gradient(Gradient {
+        colors: Cow::Borrowed(&[[1.000, 0.612, 0.867], [0.976, 0.29, 0.514]]),
+        radius: 270.0,
+    }),
+    Color::Gradient(Gradient {
+        colors: Cow::Borrowed(&[[0.584, 0.769, 0.988], [0.063, 0.165, 0.298]]),
+        radius: 270.0,
+    }),
+];
 
 pub fn config() -> (Config, HashMap<String, String>) {
     let mut displays = HashMap::new();
